// frontend/src/services/OpenApi/models/OpeningDetailsStockingSpeciesDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CodeDescriptionDto } from './CodeDescriptionDto';
export type OpeningDetailsStockingSpeciesDto = {
    layer: string | null;
    species: CodeDescriptionDto;
    minHeight: number | null;
};